backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
asyncpg>=0.29.0
aiomqtt>=2.0.0
PyJWT[crypto]>=2.8.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
